Some styling fixes
